feat (in progress) config file + cgi handle (not tested at this point)
just kind of "work in progress commit" because I just finished to write
the  config.hpp file with all classes,
A whole lot is happening in there, the main thing to understand is
that every layer of the configuration of the server get his own class

As what we call http in fact is a layer of protocols.
data wrapping data wrapping data etc etc. commentary will be added after .cpp
is written down.

and in case of wandering, yeah some one helped me a lot with that.
I am tired of the common core

josh